Javadoc

+++ src/main/java/com/mahanko/finalproject/model/dao/impl/MenuSectionDaoImpl.java
@@ -44,11 +44,11 @@ public class MenuSectionDaoImpl implements MenuSectionDao {
     }
 
     @Override
-    public Optional<MenuSection> findById(long id) throws DaoException {
+    public Optional<MenuSection> findById(Integer id) throws DaoException {
         Optional<MenuSection> sectionOptional = Optional.empty();
         try (Connection connection = ConnectionPool.getInstance().getConnection();
              PreparedStatement statement = connection.prepareStatement(SELECT_SECTION_BY_ID)) {
-            statement.setInt(1, (int) id);
+            statement.setInt(1, id);
             try (ResultSet resultSet = statement.executeQuery()) {
                 if (resultSet.next()) {
                     CustomRowMapper<MenuSection> mapper = new MenuSectionRowMapper();
